consensus: Move into internal
