+++ internal/consensus/errors.go
