+++ internal/consensus/metrics.gen.go
