+++ internal/consensus/wal_fuzz.go
